+++ Camera.h
@@ -41,7 +41,7 @@ public:
     inline const Vector3 & eye() const      {return m_eye;}
     inline const Vector3 & bgColor() const  {return m_bgColor;}
 
-    Ray eyeRay(int x, int y, int imageWidth, int imageHeight) const;
+    Ray eyeRay(float x, float y, int imageWidth, int imageHeight) const;
     
     void drawGL();
